code-examples/database-access/requirements.txt
# Python dependencies for DevLake Q Dev database access
mysql-connector-python==9.4.0
pandas==2.3.2
numpy>=1.26.0
python-dateutil>=2.8.2

# Optional dependencies for enhanced functionality
requests>=2.31.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
